refactor: packages - renamed packages/utils into packages/common and updated all affected files/places
