+++ apps/api/package.json
@@ -28,7 +28,7 @@
     "@nestjs/mapped-types": "*",
     "@nestjs/platform-express": "^10.0.0",
     "@repo/trpc": "*",
-    "@repo/utils": "*",
+    "@repo/common": "*",
     "@trpc/server": "^11.1.0",
     "@types/passport-jwt": "^4.0.1",
     "bcrypt": "^5.1.1",
