+++ apps/dashboard/package.json
@@ -18,7 +18,7 @@
     "@radix-ui/react-popover": "^1.1.6",
     "@radix-ui/react-slot": "^1.2.0",
     "@repo/trpc": "*",
-    "@repo/utils": "*",
+    "@repo/common": "*",
     "@tanstack/react-query": "^5.74.3",
     "@tanstack/react-query-devtools": "^5.74.3",
     "@trpc/client": "^11.1.0",
