+++ apps/dashboard/src/components/login-page/LoginForm.tsx
@@ -9,13 +9,13 @@ import {
 import {
   INCORRECT_PASSWORD_MESSAGE,
   USER_NOT_FOUND_MESSAGE,
-} from "@repo/utils/constants";
+} from "@repo/common/constants";
 import { Link, useNavigate } from "react-router";
 import { Button } from "../ui/button";
 import { Input } from "../ui/input";
-import { SignInUserDto } from "@repo/utils/schemas";
-import { SignInUserDtoType } from "@repo/utils/types";
-import fetchJWTToken from "@repo/utils/fetchJWTToken";
+import { SignInUserDto } from "@repo/common/schemas";
+import { SignInUserDtoType } from "@repo/common/types";
+import fetchJWTToken from "@repo/common/fetchJWTToken";
 import getCallbackUrl from "@/utils/getCallbackUrl";
 import { useForm } from "react-hook-form";
 import { useQueryClient } from "@tanstack/react-query";
