+++ apps/vscode-extension/src/utils/getGlobalStateData.ts
@@ -5,7 +5,7 @@ import {
 } from "../types-schemas";
 import { SYNC_DATA_KEY } from "../constants";
 import { ZodError } from "zod";
-import { formatZodError } from "@repo/utils/formatZodIssues";
+import { formatZodError } from "@repo/common/formatZodIssues";
 import { getExtensionContext } from "../extension";
 
 const getGlobalStateData: () => Promise<GlobalStateData> = async () => {
