+++ packages/common/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/utils",
+  "name": "@repo/common",
   "version": "0.0.0",
   "private": "true",
   "main": "./dist/index.js",
