+++ packages/common/src/formatDuration.ts
