+++ packages/common/src/getPeriodResolution.ts
